requirements.txt
absl-py==1.4.0
aiocontextvars==0.2.2
appdirs==1.4.4
astor==0.8.1
baselines==2.0
beautifultable==1.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.2.2
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
Cython==0.29.33
-e git+https://github.com/Farama-Foundation/d4rl.git@71a9549f2091accff93eeff68f1f3ab2c0e0a288#egg=D4RL
dataclasses==0.8
decorator==4.4.2
dill==0.3.4
dm-env==1.5
dm-tree==0.1.8
docker-pycreds==0.4.0
fasteners==0.18
fire==0.5.0
flatbuffers==2.0.7
future==0.18.3
gast==0.2.2
gitdb==4.0.9
GitPython==3.1.18
glfw==2.5.6
glob2==0.7
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
gtimer==1.0.0b5
gym==0.15.4
h5py==3.1.0
idna==3.4
imageio==2.15.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
jax==0.2.17
jaxlib==0.1.69
joblib==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labmaze==1.0.6
llvmlite==0.36.0
loguru==0.6.0
lxml==4.9.2
Markdown==3.3.7
matplotlib==3.3.4
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
ml-collections==0.1.1
mpi4py==3.1.3
-e git+https://github.com/spitis/mrl@7c0c9fab2698d60bbb423669f933acfbb200895c#egg=mrl
mujoco-py==2.1.2.14
mypy-extensions==1.0.0
networkx==2.5.1
numba==0.53.1
numpngw==0.1.2
numpy==1.19.5
nvidia-ml-py==11.525.84
nvitop==1.0.0
oauthlib==3.2.2
offlinerlkit==0.0.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
pandas==1.1.5
pathtools==0.1.2
patsy==0.5.3
Pillow==8.4.0
progressbar2==3.55.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycparser==2.21
pyglet==1.3.2
PyOpenGL==3.1.6
pyparsing==3.0.9
pyrallis==0.3.1
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2022.7.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.27.1
requests-oauthlib==1.3.1
-e git+https://github.com/mantle2048/rlplot@4280d0860c360896370ad4a94c9bcb8068a7696e#egg=rlplot
robosuite==1.3.0
rsa==4.9
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
sentry-sdk==1.16.0
setproctitle==1.2.3
six==1.16.0
smmap==5.0.0
statsmodels==0.12.2
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.0
tensorflow==1.15.0
tensorflow-addons==0.12.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.11.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tiffile==2018.10.18
toolz==0.12.0
torch==1.8.1+cu111
torch-tb-profiler==0.4.1
torchvision==0.9.1+cu111
tqdm==4.64.1
transforms3d==0.3.1
typing-inspect==0.8.0
typing_extensions==4.1.1
urllib3==1.26.14
wandb==0.13.10
wcwidth==0.2.6
Werkzeug==2.0.3
wgcsl==2.0
wrapt==1.15.0
zipp==3.6.0
zmq==0.0.0
